Move dtype utils to `hdf5-utils.ts`
